requirements.txt
django
djangorestframework
django-filter
ipdb
mod_wsgi
mysqlclient
six
